refactor cri
- simplify NewStreamServer
- move common util to pkg/utils
